+++ CHANGELOG.md
@@ -21,6 +21,7 @@
 
 ## Dépréciations
 
+- Le flux generique "Mail sécurisé" `mailsec` est remplacé par le flux studio `ls-mailsec` #1978
 - Le flux generique "Mail sécurisé avec réponse" `mailsec-bidir` est remplacé par le flux studio `ls-mailsec-bidir` #1978
 
 # [4.0.10] - 2023-12-11